PARQUET-1690: Integer Overflow of BinaryStatistics#isSmallerThan()

+++ parquet-column/src/main/java/org/apache/parquet/column/statistics/BinaryStatistics.java
@@ -100,7 +100,7 @@ public class BinaryStatistics extends Statistics<Binary> {
 
   @Override
   public boolean isSmallerThan(long size) {
-    return !hasNonNullValue() || ((min.length() + max.length()) < size);
+    return !hasNonNullValue() || (((long)min.length() + (long)max.length()) < size);
   }
 
   public boolean isSmallerThanWithTruncation(long size, int truncationLength) {
@@ -111,7 +111,7 @@ public class BinaryStatistics extends Statistics<Binary> {
     int minTruncateLength = Math.min(min.length(), truncationLength);
     int maxTruncateLength = Math.min(max.length(), truncationLength);
 
-    return minTruncateLength + maxTruncateLength < size;
+    return (long)minTruncateLength + (long)maxTruncateLength < size;
   }
 
   /**
